Fixed AttributeError.
When my JIRA is instantiated, but init_app() is never called (maybe when
the Flask application is executed with --help) the original JIRA
instance (which I subclass) attempts to read _options. Since I haven't
called the original JIRA __init__() method yet, this instance variable
does not exist.

Setting the variable to an empty dict in my __init__() to fix this.